add station entrance (stops.location_type = 2)  [proposal revisions]
Added note that our understanding of entrances is that they are also
used for exiting.

response to @mmorang comments
1. In the section describing location_type=2, change stop id -> stop ID
to match the rest of the page.
2. Perhaps we should add some further descriptive text explaining the
meaning of a station entrance, like "a location where passengers can
enter a station from the street"
3. Do we need to amend the text about wheelchair_boarding being
inherited from the parent station? The same rule for regular stops
should apply to station entrances, but the text is a bit awkward. Not
sure what to suggest.