Update wheelchair_boarding to include assistance
I’d like to propose an additional value in the wheelchair_boarding field: 3. 3 would denote that the path/route to the station, stop or vehicle is partially accessible, but requires additional support. 

This typically means a user can travel from the street to the platform, or from the platform to the train, but not from the street to the train, unless they solicit help from staff or someone else. This would differentiate the value from 1, which would become representative of a route that is accessible without additional support. For example, London’s TfL would map their “step-free to platform” field to this enum, which would cover the entire London Overground stations. Wheelchair-users need to call a staff hotline to get someone to help with a ramp to get on or off the train. Most of the Tokyo overground stations would also be “step-free to platform”.

This field would be valid for both trips.txt and stops.txt.