Clarify behavior for multiple fields in Alert EntitySelectors (#118)
There has been confusion in the past over the expected behavior of the EntitySelector object within Alerts, especially when multiple fields are provided within an EntitySelector (see https://groups.google.com/forum/#!topic/gtfs-realtime/jamsDygrcSk).  In particular, the GTFS-realtime .proto documentation indicates that these fields should be joined as an AND and the alert should only be applied to the entity if it matches all the provided fields. However, the GTFS-realtime documentation is somewhat ambiguous and could be interpreted as joining these fields as an OR - in other words, the alert applies to all entities that match any of the provided fields.

This proposal clarifies the GTFS-realtime documentation to align with the .proto documentation to clearly indicate that when when multiple fields are provided in an EntitySelector, they should be joined as an AND - in other words, in order for an alert to apply to an entity it must match all of the provided EntitySelector fields.