Enhance definitions for cable tram and aerial lift
Term "cable car" is used in different sense in various countries. We also want to avoid misinterpretation of the word "gondola".

US: "cable car" has cable below the vehicle and runs on rails. That is called "cable tram" outside US. Example: cable car in San Francisco.

UK: "cable car" has cable above the vehicle and uses it for support and propulsion; it has no rails. That is called "aerial lift" in US. Example: Singapore Cable Car.

Main subtypes of aerial lift are:
  a) aerial tramway - usually larger; one or two stationary ropes for support while a third moving rope provides propulsion. Examples: Roosevelt Island Tramway (New York) and Portland Aerial Tram
  b) gondola lift - usually smaller; supported and propelled by cables from above. Examples: Singapore Cable Car, Telefèric de Montjuïc.

Gondola lift is not related to gondolas in Venice.