change stops.platform_code description
The main goal of GTFS-pathways is to route anybody through a station, with a focus on step-free accessibility. (For information, the GTFS-Vehicles proposal shares the same goal but for vehicles; see the GTFS-Vehicles document). In order to route people, GTFS-Stations & GTFS-Entrances extensions are both required. The proposal is to use the field platform_code to represent the stops/platforms (location_type=0) and entrances (location_type=2) in a station in both extensions since they have similar characteristics. This field represents the identifier for a platform within a station (location_type=0) or an entrance (location_type=2) which matches the signage in the station and that is communicated to help orient the rider.