Clarification: Trip schedule_relationship and start_time for UNSCHEDULED trips
Per the discussion on the Google Group at https://groups.google.com/g/gtfs-realtime/c/ujeIp2UTnCs, TripDescriptor start_time can refer to UNSCHEDULED trips as well, so it shouldn't reference a "schedule" in it's global definition.

This change removes the word "schedule" from the initial definition and also adds an explicit reference to the UNSCHEDULED enumeration to further clarify it's use. Further clarifications surrounding TripDescriptor schedule_relationship have been added as well to make it clear that SCHEDULED is the default.